requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.2
scikit-learn==1.3.2
numpy==1.26.1
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.2.0
pyarrow==15.0.0
openpyxl==3.1.2
requests==2.31.0